file.update-time: accept int64 seconds
The primitive declared seconds-since-epoch as 'int', which on 32-bit
hosts only accepts Smi values. Smi max on 32-bit is 2^30 - 1 = ~1.07e9,
but current Unix epoch seconds are ~1.78e9, so Time.now.s-since-epoch
is a LargeInteger and the primitive rejected it with OUT_OF_RANGE.

This was uncovered by the new toit-lockfile dependency in pkg, whose
background update task calls 'file.update-time path --modification=
Time.now' every interval. We didn't notice earlier because nothing in
our own code exercised update-time with a current Time on 32-bit; the
pkg-host package has its own update-time-test, but we don't run that
package's tests in our CI.

+++ src/primitive_file_win.cc
@@ -464,7 +464,7 @@ PRIMITIVE(stat) {
 const long kToitNow = (1l << 30) - 1l;  // Special value for now.
 const long kToitNoTime = (1l << 30) - 2l;  // Special value for no time change.
 
-static void fill_time(FILETIME* time, int s, int ns, const FILETIME& current_time) {
+static void fill_time(FILETIME* time, int64 s, int ns, const FILETIME& current_time) {
   if (ns == kToitNoTime) {
     time->dwLowDateTime = 0;
     time->dwHighDateTime = 0;
@@ -483,7 +483,9 @@ static void fill_time(FILETIME* time, int s, int ns, const FILETIME& current_tim
 }
 
 PRIMITIVE(update_times) {
-  ARGS(WindowsPath, path, int, atime_sec, int, atime_nsec, int, mtime_sec, int, mtime_nsec);
+  // Seconds-since-epoch exceed Smi range on 32-bit hosts (Smi max is ~1.07e9,
+  // current epoch is ~1.78e9), so we take them as int64.
+  ARGS(WindowsPath, path, int64, atime_sec, int, atime_nsec, int64, mtime_sec, int, mtime_nsec);
   if (atime_nsec == kToitNow || atime_nsec == kToitNoTime) {
     if (atime_sec != 0) FAIL(INVALID_ARGUMENT);
   }
